feat(dho): Wiki tab for space memory, replace Artifact route
Adds /wiki with SpaceMemorySection; nav label Wiki and Library icon; getActiveTabFromPath uses wiki. /artifact permanently redirects to /wiki. i18n: Common.Wiki, CoherenceTab.wiki. Docs: dho-wiki-tab-spec, follow-up is rich media on same APIs.
