--- apps/web/src/app/[lang]/dho/[id]/@tab/artifact/constants.ts
@@ -1,5 +1,0 @@
-import { Locale } from '@hypha-platform/i18n';
-
-export const getDhoPathArtifact = (lang: Locale, id: string) => {
-  return `/${lang}/dho/${id}/artifact`;
-};
