+++ apps/web/src/app/[lang]/dho/[id]/@tab/artifact/page.tsx
@@ -1,10 +1,4 @@
-import { SpaceMemorySection } from '@hypha-platform/epics';
-import { getDhoPathAgreements } from '../agreements/constants';
-import { getDhoPathCoherence } from '../coherence/constants';
-import {
-  getEnableSpaceMemory,
-  getEnableCoherence,
-} from '@hypha-platform/feature-flags';
+import { getDhoPathWiki } from '../wiki/constants';
 import { Locale } from '@hypha-platform/i18n';
 import { redirect } from 'next/navigation';
 
@@ -12,27 +6,8 @@ type PageProps = {
   params: Promise<{ lang: Locale; id: string }>;
 };
 
-export default async function DhoArtifactPage(props: PageProps) {
-  const params = await props.params;
-  const { lang, id } = params;
-
-  const spaceMemoryEnabled = await getEnableSpaceMemory();
-  if (!spaceMemoryEnabled) {
-    const coherenceEnabled = await getEnableCoherence();
-    redirect(
-      coherenceEnabled
-        ? getDhoPathCoherence(lang, id)
-        : getDhoPathAgreements(lang, id),
-    );
-  }
-
-  return (
-    <div className="w-full min-w-0">
-      <div className="rounded-2xl border border-border/60 bg-card/35 py-4 shadow-sm backdrop-blur-[2px] supports-[backdrop-filter]:bg-card/25 dark:bg-card/40 dark:supports-[backdrop-filter]:bg-card/30">
-        <div className="px-4 pb-4 pt-0 md:px-8">
-          <SpaceMemorySection spaceSlug={id} standalonePage />
-        </div>
-      </div>
-    </div>
-  );
+/** @deprecated Use `/wiki`. Space memory lives under Wiki. */
+export default async function DhoArtifactRedirectPage(props: PageProps) {
+  const { lang, id } = await props.params;
+  redirect(getDhoPathWiki(lang, id));
 }
