+++ apps/web/src/app/[lang]/dho/[id]/@tab/wiki/constants.ts
@@ -1,0 +1,5 @@
+import { Locale } from '@hypha-platform/i18n';
+
+export const getDhoPathWiki = (lang: Locale, id: string) => {
+  return `/${lang}/dho/${id}/wiki`;
+};
