+++ packages/i18n/src/messages/de.json
@@ -44,6 +44,7 @@
     "Coherence": "Kohärenz",
     "Signals": "Signale",
     "Artifact": "Artefakt",
+    "Wiki": "Wiki",
     "uploadLeadImage": {
       "cropTitle": "Ein Bild hochladen",
       "cropDescription": "Passen Sie den Ausschnitt an und speichern Sie. Ziehen Sie zum Verschieben und zoomen Sie, um den Bannerbereich auszufüllen (JPEG, PNG oder WebP — max. 4 MB).",
@@ -1982,6 +1983,7 @@
     "callPaneResizeSpeakerShare": "Sprecher*in und Bildschirmfreigabe skalieren"
   },
   "CoherenceTab": {
+    "wiki": "Wiki",
     "artifact": "Artefakt",
     "signals": "Signale",
     "signInToSee": "Bitte melden Sie sich an, um Signale, Artefakte und Gespräche zu sehen",
