+++ packages/i18n/src/messages/en.json
@@ -44,6 +44,7 @@
     "Coherence": "Coherence",
     "Signals": "Signals",
     "Artifact": "Artifact",
+    "Wiki": "Wiki",
     "uploadLeadImage": {
       "cropTitle": "Upload an image",
       "cropDescription": "Adjust how your image is framed, then save. Drag to reposition and use zoom to fit the banner area (JPEG, PNG, or WebP — max 4 MB).",
@@ -2013,6 +2014,7 @@
     "callPaneResizeSpeakerShare": "Resize speaker and screen share"
   },
   "CoherenceTab": {
+    "wiki": "Wiki",
     "artifact": "Artifact",
     "signals": "Signals",
     "signInToSee": "Please, sign in to see signals, artifacts, and conversations",
