+++ packages/i18n/src/messages/es.json
@@ -44,6 +44,7 @@
     "Coherence": "Coherencia",
     "Signals": "Señales",
     "Artifact": "Artefacto",
+    "Wiki": "Wiki",
     "uploadLeadImage": {
       "cropTitle": "Subir una imagen",
       "cropDescription": "Ajusta el encuadre y guarda. Arrastra para reposicionar y usa el zoom para el área del banner (JPEG, PNG o WebP — máx. 4 MB).",
@@ -1981,6 +1982,7 @@
     "callPaneResizeSpeakerShare": "Redimensionar orador y pantalla"
   },
   "CoherenceTab": {
+    "wiki": "Wiki",
     "artifact": "Artefacto",
     "signals": "Señales",
     "signInToSee": "Por favor, inicie sesión para ver señales, artefactos y conversaciones",
