+++ packages/i18n/src/messages/fr.json
@@ -44,6 +44,7 @@
     "Coherence": "Cohérence",
     "Signals": "Signaux",
     "Artifact": "Artefact",
+    "Wiki": "Wiki",
     "uploadLeadImage": {
       "cropTitle": "Téléverser une image",
       "cropDescription": "Ajustez le cadrage, puis enregistrez. Faites glisser pour repositionner et zoomez pour remplir la zone de bannière (JPEG, PNG ou WebP — max. 4 Mo).",
@@ -1981,6 +1982,7 @@
     "callPaneResizeSpeakerShare": "Redimensionner l’orateur et le partage d’écran"
   },
   "CoherenceTab": {
+    "wiki": "Wiki",
     "artifact": "Artefact",
     "signals": "Signaux",
     "signInToSee": "Veuillez vous connecter pour voir les signaux, les artefacts et les conversations",
