+++ packages/i18n/src/messages/pt.json
@@ -44,6 +44,7 @@
     "Coherence": "Coerência",
     "Signals": "Sinais",
     "Artifact": "Artefato",
+    "Wiki": "Wiki",
     "uploadLeadImage": {
       "cropTitle": "Carregar uma imagem",
       "cropDescription": "Ajuste o enquadramento e salve. Arraste para reposicionar e use o zoom para a área do banner (JPEG, PNG ou WebP — máx. 4 MB).",
@@ -1981,6 +1982,7 @@
     "callPaneResizeSpeakerShare": "Redimensionar orador e partilha de ecrã"
   },
   "CoherenceTab": {
+    "wiki": "Wiki",
     "artifact": "Artefato",
     "signals": "Sinais",
     "signInToSee": "Por favor, faça login para ver sinais, artefatos e conversas",
